[Passpoint] Changes required for bouncycastle update
Latest bouncycastle changed the return type of
ASN1InputStream#readTaggedObject from DERTaggedObject to
DLTaggedObject. Since the code verified that the returned object
is an instance of DERTaggedObject, Passpoint and the unit test
failed.

Updated the logic to check instanceof either classes because
the Wi-Fi module needs to be compatible with both versions of
bouncycastle.

Bug: 181193961
Test: atest ServiceProviderVerifierTest
Change-Id: I5c8af8f003c8ef4e442cfb581beb9a9b091219df