Schedule Wifi periodic scan with exact intervals.
Currently the periodic scan is scheduled by calling AlarmManager.set,
the precision is coarse and not meet the scan interval expectation.
Change to post handler message for accurate scan interval scheduling.
The periodic scan is only scheduled when screen is on, will cancel the
scheduled scan when display is turned off.

Bug: 197671361
Test: Wifi disconnected and scan wifi when screen is on, check wifi dump for scan
intervals.

Change-Id: If19f67de8ed149e66893faf0bc12d57a37dbc7d0