wifi: log a warning and continue for mismatch config keys
The key of a WPA3 Enterprise configuration is "SSID"WPA_EAP,
but not "SSID"WPA3_EAP because there is no WPA3_EAP type string
before. This results in mismatch key issues after upgrading
to S.

The key does not belong to SDK, so we can't expect this to be the same.
For a mismatch key, just log a warning and continue.

Bug: 197879709
Test: atest FrameworksWifiTests
Change-Id: Id6ab07bd9b734dbae85d656a24494b33e2895c79