Do not require BSSID match for WifiNetworkSpecifier dialogue bypass
Problems with the current dialogue bypass model
- Needs location access: Very often apps only know the SSID but not the BSSID of the network they want to connect. The only way to get the BSSID is to do wifi scanning but this will require location access.
- MAC randomization of BSSIDs: The MAC address of the peer device could change depending on its implementation of MAC randomization. But the current code only allows the user dialogue to get bypassed if the BSSID had been observed before, so it will frustrate the user if a dialogue shows up to ask for permission to the same SSID each time the BSSID changes.

This CL makes it so that for non-open networks protected by credentials,
we only check if the SSID has been approved before. Only check for both
SSID and BSSID for open network because they are not protected and
therefore open for abuse.

Bug: 197776854
Test: atest com.android.server.wifi
Test: Manually verified with shell command that a wpa network can bypass
dialogue with just SSID match, and open network cannot bypass.

Change-Id: I3c95f1e6da858e56062bad28345b6e0327a0159d