wifi: Configure WPA2 only mode when 64-octet Hex PSK is set
When using WPA3 (SAE), all passwords in all lengths are strings, but
when using WPA2, there is a distinction between 8-63 octets that go
through BDKDF2 function, and 64-octets that are assumed to be the
output of it. BDKDF2 is not applicable to SAE and to prevent interop
issues with APs when 64-octet Hex PSK is configured, update the
configuration to use WPA2 only.

Bug: 198177036
Test: atest FrameworksWifiTests
Change-Id: I138f77ef085f13109274956293619ff76783388f