Do not update #status of current WifiConfiguration
WifiBlocklistMonitor sets the #status of a config to ENABLED whenever
the network selection status changes to enabled. Since a currently
connected network may have their network selection status enabled on
validation, the #status of this network should remain CURRENT and not
be changed to ENABLED. Otherwise, the wrong #status value will be used
when removing excess networks, and the current network may be removed
instead of a non-current network.

Bug: 202358135
Test: atest WifiBlocklistMonitorTest, manually hard code a max limit on
the number of configs and verify connected network
is not removed when adding networks past the limit.

Change-Id: I8777dd659ec2c73a8a6e6e4ac0526b1dd5eb63f7