Stop secondary CMMs when disabling wifi
Currently when wifi scanning is on we do not explicitly disable
secondary CMMs when wifi is toggled off.
Though WifiNetworkFactory will turn off the local only
secondary when the primary is gone, other usecases like the secondary
long lived and secondary trasient will not turned off, and could react
in strange ways to the primary CMM being switched to scan only mode.

This CL ensures that all secondary CMMs will get stopped when wifi is
disabled but wifi scanning is enabled.

Bug: 202809243
Test: atest com.android.server.wifi
Change-Id: I71ed2a890ab445413f21e48b806e2f52dfba4877