Fix WifiInfo#getHiddenSsid()
Before Android O, saved networks were stored by wpa_supplicant and were
considered hidden if the SSID was encoded as all-zeroes. After the
WifiConfigStore refactor in Android O, wpa_supplicant is not used to
store saved networks anymore, and whether a network is hidden or not
could be queried from WifiConfiguration. Therefore we should convert
WifiInfo#getHiddenSsid() to use the flag in WifiConfiguration.

Bug: 136480579
Test: atest ClientModeImplTest
Change-Id: I43de7ddd08b2dbe4fe62cd7feb33911ab8d94817