Catch BackendBusyException to prevent crash
BackendBusyException is a transient failure from the keystore. When this
exception is caught we will retry to get the hash function, and if the
retry also fails wifi will proceed to connect with last used randomized
MAC for the particular network. If the network has never been connected
before a new randomized MAC will get generated.

Bug: 202983511
Test: atest com.android.server.wifi
Change-Id: I4c2bea1cbafce6abda9d285911e59e9de9d699ab