Upgrade to permanent wrong password failure
Currently we only disable a network permanently for wrong password and
show the user the notification if the network had never been connected
before. This leads to problem that in a certain edge case we will
infinitely retry on a network that has a changed password. Also, another
problem is the user does not get to see the notification to remind them
to check the password, since that notification is only tied to the
permanent wrong password disable reason.

This CL will make sure that even a previously connected network will
eventually get permanently disabled for wrong password if it
consecutively fail to connect.

Bug: 203017878
Bug: 199121283
Test: atest com.android.server.wifi
Test: manually verified

Change-Id: I8e9687cae8d6b28796604d06dacb4a0348bcdc02