[OsuLogin] Handle web pages with bad SSL result
In case of TLS authentication failure result on the URL provided
by the OSU server, automatically cancel the operation, close the
webview instance, disconnect from the network, and display an
error toast - This is in line with the existing behavior of bad
URLs (e.g. failed to load).

Bug: 204478444
Test: Mock the URL to https://expired.badssl.com/ and other pages
with expected failures and confirm correct behavior.

Change-Id: I4b1f41a8c526147aff5ff3bed258ca2987cd8bb7