[WifiKeyStore] Fix update flow for app owned certs
When an app creates an Enterprise connection with certs, the
material is then moved to KeyStore and only the aliases are kept.
When then manually updating the connection from Settings, the
certs are not available, which caused a bug in the logic that
removed all certificates associated with this configuration.

Bug: 205300194
Test: atest WifiKeyStoreTest
Test: Manual test with TestPDC app
Change-Id: I240c1b1766ba1dcbe23e68578ecc39b2a061ace9