wifi: Wi-Fi data usage per configuration (Wifi part)
1. Create a new key to present the current wifi network and store it in
WifiInfo. The key will be used to be a key for Wi-Fi data usage bucket.

2. Based on the current key design, we don't need to persist the key to
DB since it can be generated by others persist data. Create a new API in
WifiConfiguration to provide all possible keys for the caller who need
to get the key to query Wi-Fi data usage.

Bug: 197520752
CTS-Coverage-Bug: 197520752
Test: atest -c FrameworksWifiTests
Test: atest -c FrameworksWifiApiTests
Change-Id: I2fc79c0b059bf968e49eedddc806d7207686b363