Trigger a scan when connected score breaches
Currently framework scan only happen periodically, so it's possible that
wifi degrades and disconnect before a scan even gets to happen.

This CL makes it possible for OEMs to configure for additional scan
being triggered when the AOSP score falls below a configurable
threshold. There is also another parameter to control the minimum period
of these additional scans.

Bug: 207695170
Bug: 207307906
Test: atest com.android.server.wifi
Change-Id: If6b36a7253bf4ce079e7cd5088e60be335b7fa8b