Toggling wifi on when only secondary CMM available should remake primary
The current issue is that when there are ANY ClientModeManagers up,
ActiveModeWarden will try to switch roles of existing primary or scan
only CMM when wifi gets toggled instead of starting a new one. But,
sometimes wifi gets into a bad state where only secondary CMMs are up,
and wifi toggle on essentially becomes no-op since there are no valid
targets to switch role.

This CL fixes the above issue by making ActiveModeWarden check to make
sure there actually exists a primary or scan only mode CMM when wifi is
toggled. If there are any, then proceed with existing flow to switch
them to the right role. If there are none, then start a new CMM with the
correct role.

Bug: 207426213
Test: atest com.android.server.wifi

Change-Id: I26662af4a09939eef9d2628f6e5ab7973bfc3b87