wifi: fix duplicate saved networks on R
For R or older, when auto-upgrade is enabled, auto-upgrade type is
matched implictly and not saved. With mulit-type configurations,
it results in duplicate saved networks.

When auto-upgrade is disable and an auto-upgrade network is
connected, the returning network id cannot be found correctly
in Settings because auto-upgrade type is not returned.

As Settings/WifiTrackerLib does the type matching by itself,
wifi service does not need to convert multi-type configs to
single-type configs. Without spliting configs, there will no
duplicate save network anymore.

Bug: 206676523
Test: atest FrameworksWifiTests
      connect to a WPA2/WPA3 network and there is only one entry in the save netwok page
      connect to a Open/OWE network and there is only one entry in the save netwok page
Change-Id: I4b9eade853eae93df15c8112e641e52f84d63fce