Implement P2P Invitation Received dialog in WifiDialogActivity
Create a WifiDialogActivity to host dialogs in the WifiDialog app,
and implement the P2P Invitation Received dialog. Create a
WifiDialogManager class to manage launching dialogs and receiving
replies in the framework. This CL does not route the dialog reply back
to the framework -- that will be done in a future CL.

Bug: 209032090
Test: verify `adb shell cmd wifi launch-dialog-p2p-invitation-received`
brings up a dialog.

Change-Id: I60e6d9d8e71b38d3604dfaeba0a1bcbddf9a3c2c