Do not allow fallback to single STA for local-only when STA+STA is supported
Local-only STA connection (aka WifiNetworkSpecifier) can replace the
primary internet providing STA - that is the use-case it was created for
(with user approval of course). However, when STA+STA is supported then
there's no need for this replacement.

Such replacement however can still happen if the other interfaces cannot
be removed. E.g. when STA+(STA||P2P) is supported and the local-only STA
cannot remove the P2P (e.g. due to priority) it now will replace the
primary STA - effectively a priority inversion (since it is the P2P
which is now out-prioritizing the primary STA - which it couldn't do on
its own).

Bug: 212587062
Test: atest com.android.server.wifi
Change-Id: I475f8c152c45dd549463a927f8120dd5b933622f