Check networkId to determine new connection
Previously the SSID of the disconnected network is compared with the
SSID of the connecting network to determine whether the disconnect is
from the old network disconnecting or the new network failing to
connect. The issue was that SSID match is not 100% reliable since there
could be networks with same SSID but different security types, and also
Passpoint.

This CL switches the SSID match to networkId match instead.

Bug: 212358864
Bug: 211950301
Test: atest com.android.server.wifi
Change-Id: I17e9e6c25c62b1a4684a105a138833c69ed92cef