Revert "Check networkId to determine new connection"
This reverts commit c64cee70324666dcfad3d57cccf8f71cb476e8ac.

Reason for revert: As per b/212358864 comment#17 this fix does not work when factory MAC is being used for both networks due to some difference in the order of execution of code. Meanwhile I have sent another fix for the partner to verify. The plan is to revert this first, and then submit the other CL once they verify it works.

Change-Id: I6bcbf7a186b9afb27d6d912a95bf8fe8cfd20279