Revoke WifiNetworkSpecifier Ux bypass in some cases
A previous CL on this stack does not allow a STA+STA to
fallback to single STA: thus not guaranteeing that WifiNetworkSpecifier
requests will be fullfilled.

A later CL provides a priority escalation to such CLs - making them
successfull against *other* (non-primary STA interfaces). However, this
priority escalation only works if the user approved the request this
specific time - not previously (i.e. not if the current request had an
bypass). This may be confusing to the app since they don't why the
request was denied (which could always happen) and don't know how to
force a user approval again.

In such scenarios, instead of rejecting the request, this CL revokes the
bypass (i.e. the previous approval) and asks the user to approve again
(which will provide a priority escalation). The bypass revocation is
limited to only requests which would be presented to the user. I.e. a
request which has a bypass through other means (privilege, CDM) doesn't
get its bypass revoked, doesn't get presented to the user, doesn't get a
priority escalation and may not get approved - which such apps need to
deal with no matter what.

Bug: 212587062
Test: atest com.android.server.wifi
Change-Id: Ibe3abe1faf3a016555b8b47eb7acb8fe047b7b6f