Do not destroy MBB secondary when captive portal is detected
Captive portal is always detected with failed internet validation. The
current logic is to disconnect the primary and have MakeBeforeBreak
manager upgrade the secondary STA with capport to primary automatically.
But because captive portal detection and no internet basically happen
back to back, sometimes a race happens causing the secondary STA to
disconnect before it could switch to primary.

Therefore, we need to avoid disconnecting the capport secondary due to
no internet, so that it can get correctly transitioned to primary STA.

Bug: 205093620
Test: atest com.android.server.wifi
Change-Id: Ia9cf7ec76b54d95baf483bbff331d8ed3bdc04f4