HDM: new method to query requirements to create an interface
Create a new method which returns the requirements to create a new
interface (without actually creating the interface). The return value
is:
- Null if the interface can't be created
- Empty list if the interface can be created w/o destroying any other
interfaces
- A list of interface types and worksources of interfaces that would be
destroyed in order to create this interface

The analysis is based on what we would do (i.e. a dummy/test run
'create') - that means that if the current priorities don't allow
creating the interface this method will not return a list of existing
interfaces.

Bug: 202813108
Test: atest HalDeviceManagerTest
Change-Id: Iba5055e53dfb738a54d32049115edaefc07650f7