Make wifi jars not depend statically on androidx
The androidx annotation library is about to start depending on kotlin,
which, with a static dependency means the kotlin-stdlib is getting
included in the wifi jars.

Since the annotations are only used at compile-time, move the dependency
to a classpath dep instead. Longer term, we should not be dependning
on androidx from these jars at all.

Bug: 214255490
Test: m framework-wifi.impl (with new androidx)
Change-Id: Ie73e5a6d8128fb46e2de815c77f3cc22e9169835