Screen on only update link layer stats when connected
Skip getting link layer stats when wifi is not connected. This CL
enhances the performance of the wifi thread.

Also fixes a race issue in test code. The test was using
startAutoDispatch with starts a new thread to run jobs posted to the
wifi handler thread. However, since the wifi method being tested has no
blocking calls, the method being tested actually finished before the
auto dispatcher finished it's job, causing the following assertion to be
wrong. The fix replaces the auto dispatch with "dispatchAll", which is
synchronous and won't have race issues.

Bug: 211564758
Test: atest com.android.server.wifi
Change-Id: I1f353699dbe9e225a9086ec74acdd007fa16bf4d