Return supported feature set even when Wi-Fi is OFF
When supplicant is started, save the complete list
of features retrieved from WiFi HAL and Supplicant HAL in config store.
When asked for supported features, include the complete
feature set stored in config store and the feature set
saved in interface class. This is to make sure that
framework is returning the feature set for
SoftAp only products and multi-chip products.

Bug: 206063858
Test: Confirmed that wifi_native_supported_features
entry in WifiConfigStore.xml is updated.

Change-Id: I312424e69dd8283c6c215c141033cb6d95793572