Migrate to V1_6 IfaceConcurrencyTypes in HDM
Modify HalDeviceManager to use the new V1_6 IfaceConcurrencyTypes. V1_0
IfaceTypes used for concurrency will still be supported for HAL versions
less than V1_6, but these will be converted to the new type in order to
avoid keeping concurrency logic in the old V1_0 IfaceType.

Non-concurrency usage of IfaceType will remain as-is.

Bug: 207055799
Test: atest FrameworksWifiTests
Change-Id: I52540e18bc0c439b4324f6235628cb4ab82fd57b