Remove WorkSpace from the AIDL interface
The official API provides the package name. It extracts it from the
WorkSpace - which contains additional information (e.g. UID). However,
the WorkSpace was transmitted over AIDL to the calling process - i.e.
the information was leaked across - whether or not it was actually
provided as part of the API.

Remove the WorkSpace and only send across the necessary information.

Bug: 202813108
Test: atest WifiServiceImplTest WifiManagerTest
Change-Id: Ibb9d320a52cfc2bcd21f33fb19105a760020e90e