Fix NetdWrapper concurrent modify exception
Access to mObervers need to be all contained in the wifi handler thread
to avoid ConcurrentModificationException. As of now the triggering of
callbacks is done in the wifi handler thread, but the add/remove is not.

This CL makes the add/remove also be done in the wifi handler thread.

Bug: 222409132
Test: atest com.android.server.wifi
Change-Id: If4b870c4909857d8d9f4be33260d88ad4203c718