Avoid calling WifiConfiguration copy constructor with null
Ensures that the existing config is found properly in an edge case where
the new config being updated doesn't have the allowKeyManagement field
set. Without this patch an invalid WifiConfiguration would be returned
and could lead to null pointer errors down the line.

This CL also adds annotations in WifiConfigManager for better
readibility.

Bug: 223980581
Test: atest com.android.server.wifi
Change-Id: I0fb3e5d65b96dd4529f562ef41ddf07cd48c3755