wifi: allow non-CA cert for TOFU
Some existing CA certs are not configured properly and
there is no basic constraints to indicate CA role.
To be compatible with those broken CA certs, non-CA cert
cannot be ignored. TOFU will use the first cert as server
cert as wpa_supplicant reports the cert from the top
of the chain.

Bug: 225092284
Test: atest FrameworksWifiTests
      connect to an EAP-PEAP network with TOFU option
Change-Id: Iadbf77b9e16288a1945f2112d35b94486f1cdcd8