wifi: set domain name to the common name of the server cert
The domain name suffix is matched against the common name
of the server cert itself, but not the one of Root CA cert.

The server cert will be the tail of the chain, i.e. the depth is 0,
so the common name of the depth 0 cert is set to the domain name
suffix.

Bug: 224474604
Test: - atest FrameworksWifiTests
      - connect to an EAP-PEAP network whose server cert common name is
        different from ca cert common name
      - connect to an EAP-PEAP network with intermediate CA cert
Change-Id: I471a60b2242bf0315a65c3bc48e9cd258db6baaf