wifi: run network selection for manual connection request
When there are multiple BSSes for a network, the candidate
might be different between network selection and manual
connection because manual connection always selects the first
scan result.

To have consistent result for a network with multiple
BSSes, wifi service should still do network
selection against this network when this network is
connected manually.

Bug: 225259099
Test: - atest FrameworksWifiTests
      - connect to a network when the device is connected, and ensure
        network selection is run.
      - connect to a network when the device is disconnected, and ensure
        network selection is run.
      - connect to a network with multiple BSSes and ensure network
        selection is run.
Change-Id: I73d83ea0a903280fb100310d459e81dfb03540f6