Fix HDM not checking if all excess ifaces can be deleted
HDM will check if an iface type can be deleted via
allowedToDeleteIfaceTypeForRequestedType, and if so, will select the
necessary quantity of those ifaces to be deleted. However, since that
method returns if ANY iface of that type can be deleted, there is no
guarantee that we can delete the necessary quantity.

Thus we should remove this method and always check if the requested
quantity can be deleted.

Bug: 219530168
Test: atest HalDeviceManagerTest
Change-Id: I6c5d5a4ec654959f3e0c99b2071a6caa7d872974