Fixed system recovery failure when only Soft Ap is enabled.
When Soft Ap is stopped, do not create client mode manager if the system
recovery is in progress.
Fixed the recovery failure when SoftAp enabled, on platforms without STA+AP
concurrency support. If a client mode manager is created when Soft Ap is
stopped, Soft Ap manager can not be created after recovery restarts.

Bug: 201125940
Test: atest ActiveModeWardenTest
Change-Id: I5a1508b0ffd15ea770f7e0749468b517a32a25e0