Identify new connection during authentication failure
When a new connection is started before the previous one is finished,
the supplicant will remove the previous network and potentially trigger
a AUTHENTICATION_FAILURE event depending on the stage of the previous
connection. ClientModeImpl needs to check the SSID associated with the
disconnect so make sure the statemachine does not mistakenly move into
disconnectedState if there's a new connection in progress.

Bug: 230143453
Bug: 229043820
Test: atest com.android.server.wifi
Test: manually verified by tapping on networks very fast
Change-Id: I7dd29cce62bbb463b09ba6040d4985aa8b2aa65b