wifi: propagate decorated anonymous identity to the supplicant
If firmware roaming is triggered before a framework roaming, i.e.
the decorated anonymous identity is not updated to the supplicant yet,
supplicant will use the original one and result in EAP failures.

To avoid using incorrect anonymous identity, the decorated one should be
propagated to the supplicant once it is generated.

Bug: 233283842
Test: atest FrameworksWifiTests
Change-Id: I838291e62eebdf2d2242c9e260e56b8ca91059fe