wifi: notify cc changed after setting even if the driver supported
The CC changed event might not be triggered even if the wlan driver
supported NL80211 REG changed event after setting country code.
The reason is the wlan driver might only trigger cc changed if there is
country code change (i.e. The framework set up a new country code which
is not the same as what driver maintained previously.).

So, this change triggers country code change after setting cc succeeds when the country code is the same as the last active country code.

Bug: 234204227
Bug: 233576163
Test: atest -c FrameworksWifiTests
Test: Manual test on Pixel 4a, CC changed callback works after wifi on
Change-Id: I7ecd61f0d225f09b6fd377a9806f2feec00b7b0e